+++ src/icemac/addressbook/browser/file/download.py
@@ -24,10 +24,9 @@ class Download(object):
         return DownloadResult(self.context)
 
 
+@zope.interface.implementer(zope.publisher.interfaces.http.IResult)
 class DownloadResult(object):
     """Result object for a download request."""
-
-    zope.interface.implements(zope.publisher.interfaces.http.IResult)
 
     def __init__(self, context):
         self._iter = bodyIterator(